+++ backend/src/routes/whitelist-routes.ts
@@ -35,11 +35,17 @@ const whitelistRoutes: FastifyPluginAsync = async (server) => {
         printRequest("/whitelist", request, server.log);
         const body: any = request.body;
         const query: any = request.query;
-        const address = body.params[0];
-        const policyId = body.params[1];
-        const useEp = query['useEp'] ?? body.params[2] ?? false;
-        const api_key = query['apiKey'] ?? body.params[3];
-        const chainId = query['chainId'] ?? body.params[4];
+        let address, policyId, api_key, chainId;
+        const useVp = query['useVp'] ?? false;
+        if(!useVp) {
+          address = body.params[0];
+          chainId = query['chainId'] ?? body.params[1];
+          api_key = query['apiKey'] ?? body.params[2];
+        } else {
+          address = body.params[0];
+          policyId = body.params[1];
+          api_key = query['apiKey'] ?? body.params[2];
+        }
         if (!api_key || typeof(api_key) !== "string")
           return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_API_KEY })
         let privateKey = '';
@@ -71,7 +77,7 @@ const whitelistRoutes: FastifyPluginAsync = async (server) => {
         }
         const validAddresses = address.every(ethers.utils.isAddress);
         if (!validAddresses) return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_ADDRESS_PASSSED });
-        if(useEp) {
+        if(!useVp) {
           if(!chainId || isNaN(chainId)) {
             return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_DATA });
           }
@@ -133,11 +139,18 @@ const whitelistRoutes: FastifyPluginAsync = async (server) => {
         printRequest("/removeWhitelist", request, server.log);
         const body: any = request.body;
         const query: any = request.query;
-        const address = body.params[0];
-        const policyId = body.params[1];
-        const chainId = query['chainId'] ?? body.params[2];
-        const api_key = query['apiKey'] ?? body.params[3];
-        const useEp = query['useEp'] ?? body.params[4] ?? false;
+        let address, policyId, api_key, chainId;
+        const useVp = query['useVp'] ?? false;
+        if(!useVp) {
+          address = body.params[0];
+          chainId = query['chainId'] ?? body.params[1];
+          api_key = query['apiKey'] ?? body.params[2];
+        } else {
+          address = body.params[0];
+          policyId = body.params[1];
+          chainId = query['chainId'] ?? body.params[2];
+          api_key = query['apiKey'] ?? body.params[3];
+        }
         if (!api_key || typeof(api_key) !== "string")
           return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_API_KEY })
         let privateKey = '';
@@ -173,7 +186,7 @@ const whitelistRoutes: FastifyPluginAsync = async (server) => {
         }
         const validAddresses = address.every(ethers.utils.isAddress);
           if (!validAddresses) return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_ADDRESS_PASSSED });
-        if(useEp) {
+        if(!useVp) {
           if(!chainId || isNaN(chainId)) {
             return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_DATA });
           }
@@ -237,11 +250,18 @@ const whitelistRoutes: FastifyPluginAsync = async (server) => {
         printRequest("/checkWhitelist", request, server.log);
         const body: any = request.body;
         const query: any = request.query;
-        const accountAddress = body.params[0];
-        const policyId = body.params[1];
-        const chainId = query['chainId'] ?? body.params[2];
-        const api_key = query['apiKey'] ?? body.params[3];
-        const useEp = query['useEp'] ?? body.params[4];
+        let accountAddress, policyId, api_key, chainId;
+        const useVp = query['useVp'] ?? false;
+        if(!useVp) {
+          accountAddress = body.params[0];
+          chainId = query['chainId'] ?? body.params[1];
+          api_key = query['apiKey'] ?? body.params[2];
+        } else {
+          accountAddress = body.params[0];
+          policyId = body.params[1];
+          chainId = query['chainId'] ?? body.params[2];
+          api_key = query['apiKey'] ?? body.params[3];
+        }
         if (!api_key || typeof(api_key) !== "string")
           return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_API_KEY })
         let privateKey = '';
@@ -276,7 +296,7 @@ const whitelistRoutes: FastifyPluginAsync = async (server) => {
         }
         const networkConfig = getNetworkConfig(chainId, supportedNetworks ?? '', SUPPORTED_ENTRYPOINTS.EPV_06);
         if (!networkConfig) return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.UNSUPPORTED_NETWORK });
-        if(useEp) {
+        if(!useVp) {
           if(!chainId || isNaN(chainId)) {
             return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_DATA });
           }
@@ -311,7 +331,220 @@ const whitelistRoutes: FastifyPluginAsync = async (server) => {
     }
   );
 
-  server.post("/getAllWhitelist",
+  server.post("/whitelist/v2",
+    async function (request, reply) {
+      try {
+        printRequest("/whitelist/v2", request, server.log);
+        const body: any = request.body;
+        const query: any = request.query;
+        const address = body.params[0];
+        const policyId = body.params[1];
+        const api_key = query['apiKey'] ?? body.params[2];
+        if (!api_key || typeof(api_key) !== "string")
+          return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_API_KEY })
+        let privateKey = '';
+        const apiKeyEntity: APIKey | null = await server.apiKeyRepository.findOneByApiKey(api_key);
+        if (!apiKeyEntity) return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_API_KEY })
+        if (!unsafeMode) {
+          const AWSresponse = await client.send(
+            new GetSecretValueCommand({
+              SecretId: prefixSecretId + api_key,
+            })
+          );
+          const secrets = JSON.parse(AWSresponse.SecretString ?? '{}');
+          if (!secrets['PRIVATE_KEY']) return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_API_KEY })
+          privateKey = secrets['PRIVATE_KEY'];
+        } else {
+          privateKey = decode(apiKeyEntity.privateKey, server.config.HMAC_SECRET);
+        }
+        if (!privateKey) return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_API_KEY })
+        if (
+          !Array.isArray(address) ||
+          address.length > 10
+        ) {
+          return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_DATA });
+        }
+        if (server.config.SUPPORTED_NETWORKS == '' && !SupportedNetworks) {
+          return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.UNSUPPORTED_NETWORK });
+        }
+        const validAddresses = address.every(ethers.utils.isAddress);
+        if (!validAddresses) return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_ADDRESS_PASSSED });
+        const signer = new Wallet(privateKey)
+        if (policyId) {
+          const policyRecord = await server.sponsorshipPolicyRepository.findOneById(policyId);
+          if (!policyRecord || (policyRecord?.walletAddress !== signer.address)) return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_SPONSORSHIP_POLICY_ID })
+        }
+        const existingWhitelistRecord = await server.whitelistRepository.findOneByApiKeyAndPolicyId(api_key, policyId);
+
+        if (existingWhitelistRecord) {
+          const toBeAdded: string[] = [];
+          address.filter(ele => {
+            if (!existingWhitelistRecord.addresses.includes(ele)) toBeAdded.push(ele);
+          });
+          if (toBeAdded.length < 1) return reply.code(ReturnCode.CONFLICT).send({ error: ErrorMessage.ADDRESS_ALREADY_ADDED });
+          const allAddresses = toBeAdded.concat(existingWhitelistRecord.addresses);
+          existingWhitelistRecord.addresses = allAddresses;
+          await server.whitelistRepository.updateOneById(existingWhitelistRecord);
+        } else {
+          const addWhitelistDto = {
+            apiKey: api_key,
+            addresses: address,
+            policyId: policyId ?? null,
+          }
+          await server.whitelistRepository.create(addWhitelistDto);
+        }
+        const result = { message: "Successfully whitelisted" }
+        server.log.info(result, 'Response sent: ');
+        if (body.jsonrpc)
+          return reply.code(ReturnCode.SUCCESS).send({ jsonrpc: body.jsonrpc, id: body.id, result, error: null })
+        return reply.code(ReturnCode.SUCCESS).send(result);
+      } catch (err: any) {
+        request.log.error(err);
+        if (err.name == "ResourceNotFoundException")
+          return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_API_KEY });
+        return reply.code(ReturnCode.FAILURE).send({ error: err.message ?? ErrorMessage.FAILED_TO_PROCESS })
+      }
+    }
+  );
+
+  server.post("/removeWhitelist/v2",
+    async function (request, reply) {
+      try {
+        printRequest("/removeWhitelist/v2", request, server.log);
+        const body: any = request.body;
+        const query: any = request.query;
+        const address = body.params[0];
+        const policyId = body.params[1];
+        const chainId = query['chainId'] ?? body.params[2];
+        const api_key = query['apiKey'] ?? body.params[3];
+        if (!api_key || typeof(api_key) !== "string")
+          return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_API_KEY })
+        let privateKey = '';
+        const apiKeyEntity: APIKey | null = await server.apiKeyRepository.findOneByApiKey(api_key);
+        if (!apiKeyEntity) return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_API_KEY })
+        if (!unsafeMode) {
+          const AWSresponse = await client.send(
+            new GetSecretValueCommand({
+              SecretId: prefixSecretId + api_key,
+            })
+          );
+          const secrets = JSON.parse(AWSresponse.SecretString ?? '{}');
+          if (!secrets['PRIVATE_KEY']) return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_API_KEY })
+          privateKey = secrets['PRIVATE_KEY'];
+        } else {
+          privateKey = decode(apiKeyEntity.privateKey, server.config.HMAC_SECRET);
+        }
+        const supportedNetworks = apiKeyEntity.supportedNetworks;
+        if (!privateKey) return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_API_KEY })
+        if (
+          !Array.isArray(address) ||
+          address.length > 10 ||
+          !chainId ||
+          isNaN(chainId)
+        ) {
+          return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_DATA });
+        }
+        if (server.config.SUPPORTED_NETWORKS == '' && !SupportedNetworks) {
+          return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.UNSUPPORTED_NETWORK });
+        }
+        const networkConfig = getNetworkConfig(chainId, supportedNetworks ?? '', SUPPORTED_ENTRYPOINTS.EPV_07);
+        if (!networkConfig) return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.UNSUPPORTED_NETWORK });
+        const validAddresses = address.every(ethers.utils.isAddress);
+        if (!validAddresses) return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_ADDRESS_PASSSED });
+        const existingWhitelistRecord = await server.whitelistRepository.findOneByApiKeyAndPolicyId(api_key, policyId);
+
+
+        if (existingWhitelistRecord) {
+          const toBeRemoved: string[] = [];
+          address.filter(ele => {
+            if (existingWhitelistRecord.addresses.includes(ele)) {
+              toBeRemoved.push(ele);
+              existingWhitelistRecord.addresses.splice(existingWhitelistRecord.addresses.indexOf(ele), 1);
+            }
+          });
+          if (toBeRemoved.length < 1) return reply.code(ReturnCode.CONFLICT).send({ error: ErrorMessage.ADDRESS_NOT_WHITELISTED });
+
+          if (existingWhitelistRecord.addresses.length < 1) await server.whitelistRepository.deleteById(existingWhitelistRecord.id);
+          else await server.whitelistRepository.updateOneById(existingWhitelistRecord);
+        } else {
+          throw new Error(ErrorMessage.NO_WHITELIST_FOUND);
+        }
+        const result = { message: "Successfully removed whitelisted addresses" }
+        server.log.info(result, 'Response sent: ');
+        if (body.jsonrpc)
+          return reply.code(ReturnCode.SUCCESS).send({ jsonrpc: body.jsonrpc, id: body.id, result, error: null })
+        return reply.code(ReturnCode.SUCCESS).send(result);
+      } catch (err: any) {
+        request.log.error(err);
+        if (err.name == "ResourceNotFoundException")
+          return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_API_KEY });
+        return reply.code(ReturnCode.FAILURE).send({ error: err.message ?? ErrorMessage.FAILED_TO_PROCESS })
+      }
+    }
+  );
+
+  server.post("/checkWhitelist/v2",
+    async function (request, reply) {
+      try {
+        printRequest("/checkWhitelist/v2", request, server.log);
+        const body: any = request.body;
+        const query: any = request.query;
+        const accountAddress = body.params[0];
+        const policyId = body.params[1];
+        const chainId = query['chainId'] ?? body.params[2];
+        const api_key = query['apiKey'] ?? body.params[3];
+        if (!api_key || typeof(api_key) !== "string")
+          return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_API_KEY })
+        let privateKey = '';
+        const apiKeyEntity: APIKey | null = await server.apiKeyRepository.findOneByApiKey(api_key);
+        if (!apiKeyEntity) return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_API_KEY })
+        if (!unsafeMode) {
+          const AWSresponse = await client.send(
+            new GetSecretValueCommand({
+              SecretId: prefixSecretId + api_key,
+            })
+          );
+          const secrets = JSON.parse(AWSresponse.SecretString ?? '{}');
+          if (!secrets['PRIVATE_KEY']) return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_API_KEY })
+          privateKey = secrets['PRIVATE_KEY'];
+        } else {
+          privateKey = decode(apiKeyEntity.privateKey, server.config.HMAC_SECRET);
+        }
+        const supportedNetworks = apiKeyEntity.supportedNetworks;
+        if (!privateKey) return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_API_KEY })
+        if (
+          !accountAddress ||
+          !ethers.utils.isAddress(accountAddress) ||
+          !chainId ||
+          isNaN(chainId)
+        ) {
+          return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_DATA });
+        }
+        if (server.config.SUPPORTED_NETWORKS == '' && !SupportedNetworks) {
+          return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.UNSUPPORTED_NETWORK });
+        }
+        const networkConfig = getNetworkConfig(chainId, supportedNetworks ?? '', SUPPORTED_ENTRYPOINTS.EPV_07);
+        if (!networkConfig) return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.UNSUPPORTED_NETWORK });
+        const existingWhitelistRecord = await server.whitelistRepository.findOneByApiKeyAndPolicyId(api_key, policyId);
+
+        if (!existingWhitelistRecord) {
+          throw new Error(ErrorMessage.NO_WHITELIST_FOUND);
+        }
+        const result = { message: existingWhitelistRecord.addresses.includes(accountAddress) ? 'Already added' : 'Not added yet' }
+        server.log.info(result, 'Response sent: ');
+        if (body.jsonrpc)
+          return reply.code(ReturnCode.SUCCESS).send({ jsonrpc: body.jsonrpc, id: body.id, result, error: null })
+        return reply.code(ReturnCode.SUCCESS).send(result);
+      } catch (err: any) {
+        request.log.error(err);
+        if (err.name == "ResourceNotFoundException")
+          return reply.code(ReturnCode.FAILURE).send({ error: ErrorMessage.INVALID_API_KEY });
+        return reply.code(ReturnCode.FAILURE).send({ error: err.message ?? ErrorMessage.FAILED_TO_PROCESS })
+      }
+    }
+  );
+
+  server.post("/getAllWhitelist/v2",
     async function (request, reply) {
       try {
         printRequest("/getAllWhitelist/v2", request, server.log);