+++ backend/CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
-## [1.8.0] - 2024-12-25
+## [2.0.1] - 2025-01-08
+### New
+- added `/whitelist/v2`, `/removeWhitelist/v2` and `/checkWhitelist/v2` endpoints back.
+- `/whitelist`, `/removeWhitelist`, `checkWhitelist`, `/deposit/v2`, `/` endpoints accept `useVp` param which when set to true uses VerifyingPaymaster contract instead of EtherspotPaymaster.
+
+
+## [2.0.0] - 2024-12-25
 ### Breaking changes
-- removed `/whitelist/v1`, `/removeWhitelist/v1`, `/checkWhitelist/v1` endpoints.
 - removed `/whitelist/v2`, `/removeWhitelist/v2`, `/checkWhitelist/v2` endpoints.
 
 ### New